tune pump controls

+++ stm32-modules/include/vacuum-module/firmware/pump_hardware.h
@@ -8,8 +8,8 @@ extern "C" {
 void pump_hardware_init(void);
 bool hw_start_pump_motor();
 bool hw_stop_pump_motor();
-void hw_set_pump_duty_cycle(uint16_t duty);
-uint16_t hw_get_pump_duty_cycle(void);
+void hw_set_pump_duty_cycle(uint8_t duty);
+uint8_t hw_get_pump_duty_cycle(void);
 bool hw_enable_pump_tach(bool enable);
 float hw_get_pump_rpm(void);
 void tach_period_overflow_callback(void);
