+++ stm32-modules/include/vacuum-module/firmware/pump_policy.hpp
@@ -15,8 +15,8 @@ class PumpPolicy {
 
     auto start_pump_motor() -> bool;
     auto stop_pump_motor() -> bool;
-    auto set_pump_duty_cycle(uint16_t duty) -> void;
-    auto get_pump_duty_cycle() -> uint16_t;
+    auto set_pump_duty_cycle(uint8_t duty) -> void;
+    auto get_pump_duty_cycle() -> uint8_t;
     auto enable_pump_tach(bool enable) -> bool;
     auto get_pump_rpm() -> float;
     auto static sleep_ms(uint32_t ms) -> void;
